@@ -4,30 +4,45 @@
   "private": true,
   "scripts": {
     "start": "node_modules/react-native/packager/packager.sh",
-    "test": "jest"
+    "test": "mocha --compilers js:babel/register --require react-native-mock/mock.js --recursive __tests__/**/*.js"
   },
   "dependencies": {
     "buffer": "^3.5.4",
-    "flux": "*",
     "jsrsasign": "5.0.1",
     "keymirror": "~0.1.0",
-    "react": "^0.14.7",
-    "react-native": "0.22.2",
+    "react": "^15.1.0",
+    "react-native": "0.27.2",
     "react-native-button": "*",
-    "react-native-vector-icons": "^1.3.3",
+    "react-native-vector-icons": "^2.0.3",
+    "react-redux": "^4.4.5",
+    "redux": "^3.4.0",
+    "redux-logger": "^2.6.1",
+    "redux-saga": "^0.10.5",
+    "redux-storage": "^4.0.0",
+    "redux-storage-decorator-filter": "^1.1.4",
+    "redux-storage-engine-reactnativeasyncstorage": "^1.0.0",
     "underscore": "*",
     "url": "^0.11.0",
     "url-join": "*"
   },
   "devDependencies": {
-    "jest-cli": "0.5.7",
+    "babel-core": "^6.9.1",
     "babel-eslint": "4.1.1",
+    "babel-preset-react-native": "^1.9.0",
+    "chai": "^3.5.0",
+    "enzyme": "^2.3.0",
     "eslint": "1.3.1",
-    "eslint-plugin-react": "3.3.1"
+    "eslint-plugin-react": "3.3.1",
+    "jest-cli": "~11",
+    "mocha": "^2.5.3",
+    "react": "^15.1.0",
+    "react-addons-test-utils": "^15.1.0",
+    "react-dom": "^15.1.0",
+    "react-native-mock": "^0.2.3"
   },
   "lint": "eslint . --ext .js --ext .jsx",
   "jest": {
-    "scriptPreprocessor": "node_modules/react-native/jestSupport/scriptPreprocess.js",
+    "scriptPreprocessor": "node_modules/react-native/jestSupport/preprocessor.js",
     "setupEnvScriptFile": "node_modules/react-native/jestSupport/env.js",
     "testPathIgnorePatterns": [
       "/node_modules/",
@@ -39,6 +54,7 @@
     "unmockedModulePathPatterns": [
       "promise",
       "source-map"
-    ]
+    ],
+    "automock": false
   }
 }
