@@ -1,23 +1,8 @@
 'use strict'
-
-var React = require('react-native');
-var NativeGlobals = React.NativeModules.TMGlobals;
 var Geo = require('./Geo')
 var urljoin = require('url-join');
 var jsrassign = require('jsrsasign');
-var UserStore = require('../stores/UserStore');
 
-var apiHost;
-switch(NativeGlobals.buildType){
-    case 'adhoc':
-        apiHost = 'http://api-staging.totem-app.com';
-        break;
-    case 'appstore':
-        apiHost = 'http://api.totem-app.com';
-        break;
-    default:
-        apiHost = 'http://localhost:3000';
-}
 
 var DEFAULT_HEADERS = {
     'Accept': 'application/json',
@@ -26,14 +11,14 @@ var DEFAULT_HEADERS = {
 
 class TotemApi{
 
-    async placesNearby(lon, lat){
+    static async placesNearby(user, lon, lat){
         var params =  {lon: lon, lat: lat};
-        var encodedParams = await this._encodeJWT(params);
-        var response = await fetch(urljoin(apiHost, '/api/v1/places/nearby.json', `?jwt=${encodedParams}`));
+        var encodedParams = TotemApi._encodeJWT(user, params);
+        var response = await fetch(urljoin(TotemApi.apiHost(), '/api/v1/places/nearby.json', `?jwt=${encodedParams}`));
         return response.json();
     }
 
-    async placeCreate(name, category_id, lon, lat){
+    static async placeCreate(user, name, category_id, lon, lat){
 
         var placeCreateParams = {
             location: Geo.jsonFromPoint(lon, lat),
@@ -42,24 +27,44 @@ class TotemApi{
               category_id: category_id}
         };
 
-        var response = await this._post('/api/v1/places.json', placeCreateParams);
+        var jwtToken = TotemApi._encodeJWT(user, placeCreateParams);
+
+        var response = await TotemApi._post('/api/v1/places.json', jwtToken);
         return response.json();
     }
-    async visitCreate(place_id, lon, lat){
+    static async visitCreate(user, place_id, lon, lat){
 
         var visitCreateParams = { visit: {place_id: place_id, location: Geo.jsonFromPoint(lon, lat)} };
+        var jwtToken = TotemApi._encodeJWT(user, visitCreateParams);
 
-        var response = await this._post(urljoin('/api/v1/places', place_id, 'visits'), visitCreateParams);
+        var response = await TotemApi._post(urljoin('/api/v1/places', place_id, 'visits'), jwtToken);
         return response.json();
     }
 
-    userSessionUrl(){
-        return urljoin(apiHost, '/users/sign_up')
+    static userSessionUrl(){
+        return urljoin(TotemApi.apiHost(), '/users/sign_up')
     }
 
-    async _encodeJWT(params){
+    static apiHost(){
+        var ReactNative = require('react-native');
+        var NativeGlobals = ReactNative.NativeModules.TMGlobals;
+        var apiHost;
+        switch(NativeGlobals.buildType){
+            case 'adhoc':
+                apiHost = 'http://api-staging.totem-app.com';
+            break;
+            case 'appstore':
+                apiHost = 'http://api.totem-app.com';
+            break;
+            default:
+                apiHost = 'http://localhost:3000';
+        }
+        return apiHost;
 
-        var user = await UserStore.getAsync();
+    }
+
+    static _encodeJWT(user, params){
+
         params['public_token'] =  user.public_token;
 
         var header = JSON.stringify({alg: 'HS256', typ: 'JWT'});
@@ -68,18 +73,16 @@ class TotemApi{
         return jsrassign.jws.JWS.sign('HS256', header, payload, user.private_token);
     }
 
-    async _post(url, params){
-        var jwtToken = await this._encodeJWT(params);
-
+    static async _post(url, jwtToken){
         var fetchOptions = {
             method: 'POST',
             headers: DEFAULT_HEADERS,
             body: JSON.stringify({jwt: jwtToken})
         };
-        return fetch(urljoin(apiHost, url), fetchOptions);
+        return fetch(urljoin(TotemApi.apiHost(), url), fetchOptions);
     }
 
 
 }
 
-module.exports = new TotemApi();
+module.exports = TotemApi;
